requirements.txt
asgiref==3.2.10
<<<<<<< HEAD
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
gcloud==0.18.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
jwcrypto==0.8
oauth2client==4.1.3
Pillow==7.2.0
protobuf==3.15.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.4.3
python-decouple==3.4
python-jwt==3.3.0
pytz==2020.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.4
whitenoise==5.2.0
=======
Django==3.1
django-crispy-forms==1.9.2
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
>>>>>>> 2960a3c15d128869b5ae620a497dde33bd7d88e9
